requirements.txt
Django
django-cors-headers
django-filter
djangorestframework
Pillow
requests
sqlparse
python-dotenv